requirements.txt
SimpleITK
numpy
timm==0.9.8
h5py
pandas
PyYAML
opencv-python
matplotlib
scikit-learn
scipy
tqdm
openslide-python
tensorboardX
addict
nystrom-attention==0.0.11
nmslib
torch_geometric
networkx